feat: Add PoissonEloEstimator (#38)
Rating system based on Poisson regression, i.e. the rating system
is equivalent to Poisson regression with stochastic gradient
descent. The outcome should be count data though can be used for
binary outcomes in cases where we are interested in the relative
risks.

To define this we:
- introduces PoissonRegression class
- add ClassifierRatingSystemMixin and RegressionRatingSystemMixin
  which inherit from RatingSystemMixin
- add a beta parameter to the abstract Model class
- add an abstract base class BaseEloEstimator

Supporting documentation is added:
- examples/football.ipynb: a NB applying PoissonEloEstimator
  to Premier League football data
- docs/feature_ref/poisson.md: mkdocs page outlining the
  theory behind the implementation