feat: Support L1 and L2 regularisation (#43)
Adds Regressor.penalty. Specifies the type of penalty for
regularisation. Can be None, 'l1' or 'l2'.

Adds Model.calculate_gradient_from_expected_score.
We will often calculate the expected score for predictions and
afterwards want to update the ratings of the entities. This method
allows us to feed the pre-calculated expected score back to the
model to calculate the update step, rather than recalculating or
caching.

Change Optimizer.calculate_update_step signature.
Replace regressor_contrib with expected_score. This is a more
meaningful and natural quantity. When expected score is not None
we use that value to avoid recalculating it, and otherwise we do
recalculate the expected_score using the model.ratings dict and
the additional_regressor_values (renamed from regressor_values for
consistency).