docs: Update roadmap (#49)
Remove plan to handle correlation in scores. This is outside the
scope of the rating systems or can potentially be handled by an
appropriate choice of rating system, e.g. a Skellam rating system
rather than two Poisson systems.

We replace this with a plan to support hierarchical matches, such
as tennis which has sets and games within a match. This would
allow us to pass a dataframe at the match level but fit a rating
system at the set or game level. We could also roll the
predictions back to the match level.